// temp/frontend-true/libs/shared/src/lib/access/access.type.ts
export type Access =
    //#region EDUCATION
    // DASHBOARD
    | 'EDUCATION_DASHBOARD_ACTIVE'
    | 'EDUCATION_DASHBOARD_STUDY'
    // REPORT
    | 'EDUCATION_REPORT_STUDY'
    | 'EDUCATION_REPORT_COURSE'
    | 'EDUCATION_REPORT_INSTITUTE'
    | 'EDUCATION_REPORT_MENTOR'
    | 'EDUCATION_REPORT_PARTICIPANT'
    | 'EDUCATION_REPORT_PERSONNEL'
    // ACCESS
    | 'EDUCATION_COURSE'
    | 'EDUCATION_MENTOR'
    | 'EDUCATION_INSTITUTE'
    | 'EDUCATION_LOCATION'
    | 'EDUCATION_ACTIVE'
    | 'EDUCATION_DONE'
    | 'EDUCATION_CANCELED'
    | 'EDUCATION_UNPAID'
    | 'EDUCATION_LOG'
    | 'EDUCATION_SETTING'
    // ROLE
    | 'EDUCATION_ROLE_STUDY'
    | 'EDUCATION_ROLE_PAYMENT'
    //#endregion

    //#region KITCHEN
    // DASHBOARD
    | 'KITCHEN_DASHBOARD_CALENDAR'
    | 'KITCHEN_DASHBOARD_SERVING'
    | 'KITCHEN_DASHBOARD_INVENTORY'
    // ACCESS
    | 'KITCHEN_GROUP'
    | 'KITCHEN_GOOD'
    | 'KITCHEN_INVENTORY'
    | 'KITCHEN_RECIPE'
    | 'KITCHEN_CALENDAR'
    | 'KITCHEN_SERVING'
    | 'KITCHEN_SERVING_LOG'
    //#endregion

    //#region LABORATORY
    // DASHBOARD
    | 'LABORATORY_DASHBOARD_DAILY'
    // REPORT
    | 'LABORATORY_REPORT_AVERAGE'
    | 'LABORATORY_REPORT_CRUSHER_LOCATION'
    | 'LABORATORY_REPORT_KHATKA_LOCATION'
    | 'LABORATORY_REPORT_CRUSHER'
    | 'LABORATORY_REPORT_KHATKA'
    | 'LABORATORY_REPORT_LOAD'
    // ACCESS
    | 'LABORATORY_CARGO'
    | 'LABORATORY_PRODUCTION_CRUSHER'
    | 'LABORATORY_PRODUCTION_KHATKA'
    | 'LABORATORY_CRUSHER'
    | 'LABORATORY_KHATKA'
    | 'LABORATORY_BLAINE'
    | 'LABORATORY_DAVIS'
    | 'LABORATORY_SOLID'
    | 'LABORATORY_LOAD'
    | 'LABORATORY_MISC'
    | 'LABORATORY_SUPPLEMENTARY'
    | 'LABORATORY_LOG'
    | 'LABORATORY_SETTING'
    // ROLE
    | 'LABORATORY_ROLE_LOAD'
    | 'LABORATORY_ROLE_MISC'
    | 'LABORATORY_ROLE_SUPPLEMENTARY'
    | 'LABORATORY_ROLE_TECHNICIAN'
    //#endregion

    //#region LOAD
    // DASHBOARD
    | 'LOAD_DASHBOARD_TYPE'
    | 'LOAD_DASHBOARD_ACTIVE'
    | 'LOAD_DASHBOARD_CARGO'
    | 'LOAD_DASHBOARD_CHART'
    // REPORT
    | 'LOAD_REPORT_ACTIVE'
    | 'LOAD_REPORT_DRAFT'
    | 'LOAD_REPORT_DAILY'
    | 'LOAD_REPORT_PARTY'
    | 'LOAD_REPORT_SHIPMENT'
    | 'LOAD_REPORT_TRANSPORTER'
    | 'LOAD_REPORT_CARGO'
    | 'LOAD_REPORT_OWNER'
    | 'LOAD_REPORT_TRUCK'
    | 'LOAD_REPORT_DAILY_TRANSPORTER'
    // ACCESS
    | 'LOAD_PARTY'
    | 'LOAD_SHIPMENT'
    | 'LOAD_MISC'
    | 'LOAD_TRANSPORTER'
    | 'LOAD_CARGO'
    | 'LOAD_OWNER'
    | 'LOAD_TRUCK'
    | 'LOAD_CHECKOUT'
    | 'LOAD_SETTING'
    | 'LOAD_DATA_LOG'
    | 'LOAD_DRAFT_DAILY'
    | 'LOAD_DRAFT_ACTIVE'
    | 'LOAD_DRAFT_FINISHED'
    | 'LOAD_DRAFT_CANCELED'
    | 'LOAD_DRAFT_UPDATED'
    | 'LOAD_DRAFT_LOG'
    | 'LOAD_DRAFT_UPDATE'
    | 'LOAD_FLOW_CANCEL'
    // ROLE
    | 'LOAD_ROLE_TRAFFIC'
    | 'LOAD_ROLE_TRAFFIC_MINE'
    | 'LOAD_ROLE_WEIGHT'
    | 'LOAD_ROLE_LOADING'
    | 'LOAD_ROLE_LOADING_MINE'
    | 'LOAD_ROLE_DISCHARGE'
    //#endregion

    //#region PERSONNEL
    // REPORT
    | 'PERSONNEL_REPORT_MEMBER'
    // ACCESS
    | 'PERSONNEL_GROUP'
    | 'PERSONNEL_MEMBER'
    | 'PERSONNEL_LOCATION'
    | 'PERSONNEL_SEARCH'
    | 'PERSONNEL_STATUS'
    | 'PERSONNEL_EXPORT'
    // ROLE
    | 'PERSONNEL_ROLE_MEMBER'
    //#endregion

    //#region SUPPORT
    // ACCESS
    | 'SUPPORT_TICKET'
    | 'SUPPORT_NOTIFICATION'
    | 'SUPPORT_ALERT'
    | 'SUPPORT_SETTING'
    //#endregion

    //#region TRANSPORT
    // ACCESS
    | 'TRANSPORT_GROUP'
    | 'TRANSPORT_LOCATION'
    | 'TRANSPORT_PARKING'
    | 'TRANSPORT_IMPORT'
    | 'TRANSPORT_FINAL'
    // ROLE
    | 'TRANSPORT_ROLE_STATION'
    | 'TRANSPORT_ROLE_ROUTE'
    //#endregion

    //#region WAREHOUSE
    // ACCESS
    | 'WAREHOUSE_CATEGORY_LOG'
    | 'WAREHOUSE_STOCK_LOG'
    | 'WAREHOUSE_INVENTORY'
    | 'WAREHOUSE_EXPORT'
    | 'WAREHOUSE_SETTING'
    | 'WAREHOUSE_HELP'
    // ROLE
    | 'WAREHOUSE_CATEGORY'
    | 'WAREHOUSE_STOCK'
    | 'WAREHOUSE_DELETE';
//#endregion